backend/requirements.txt
# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# 异步和HTTP
aiohttp==3.9.1
asyncio==3.4.3
httpx==0.26.0

# 浏览器自动化
playwright==1.40.0

# 消息队列
redis==5.0.1
aioredis==2.0.1

# 消息转发
discord-webhook==1.3.0
python-telegram-bot==20.7
lark-oapi==1.2.9

# 数据库
aiosqlite==0.19.0

# 工具库
cryptography==41.0.7
Pillow==10.1.0
python-dotenv==1.0.0
pydantic-settings==2.1.0

# 日志
loguru==0.7.2

# 日期时间
python-dateutil==2.8.2
